# app_shop/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0", 
      "title": "My User Project CRUD",
      "description": "My User Project Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/orders":{
            "parameters": [
                {
                    "name": "_id",
                    "in":"query",
                    "required":true,
                    "type": "string",
                    "description": "User id"
                },
                {
                    "name": "type",
                    "in":"header",
                    "required":false,
                    "type": "string",
                    "description": "Orders info"
                }
            ],

            "get": {
                "tags": [
                    "Orders"
                ],
                "description": "Find information about all Orders by User id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                },
                "summary": "Get user Orders"
            }
        },
...
        "/cart":{
    
            "get": {
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "_id",
                        "in":"query",
                        "required":true,
                        "type": "string",
                        "description": "User id"
                    },
                    {
                        "name": "type",
                        "in":"header",
                        "required":false,
                        "type": "string",
                        "description": "Orders info"
                    }
                ],
                "description": "Find information about all Products in Cart User id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Cart"
                        }
                    }
                },
                "summary": "Get user Orders"
            }
        },
        "/products":{
            "parameters": [
                {
                    "name": "page",
                    "in":"query",
                    "required":true,
                    "type": "string",
                    "description": "Page number"
                },
                {
                    "name": "type",
                    "in":"header",
                    "required":false,
                    "type": "string",
                    "description": "Products"
                }
            ],

            "get": {
                "tags": [
                    "Cart"
                ],
                "description": "Find Products on certain page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                },
                "summary": "Products on page"
            }
        },
...
    },
    "definitions": {
        "Product": {
            "required": [
                "title",
                "prime_price",
                "price",
                "description",
                "imageUrl"
            ],
            "properties": {
                "title": {
                    "type": "string"
                },
                "prime_price": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "num_purch": {
                    "type": "number"
                },
                "num_deletions": {
                    "type": "number"
                },
                "date_add": {
                    "type": "string",
                    "format": "date"
                },
                "date_purch": {
                    "type": "string",
                    "format": "date"
                }

            }
        },
        "Products": {
            "type": "array",
            "items":{
                "$ref": "#/definitions/Product"
            }       
        },
        "Cart": {
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items":{
                        "type":"object",
                        "properties":{
                            "productId": {  
                                "$ref": "#/definitions/Product" 
                            },
                            "quantity": { 
                                "type": "number"
                            }
                        }
                    }    
                }
            }
        },
...
    }
}